backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
alembic
python-dotenv
pydantic
pydantic-settings
python-jose
passlib[bcrypt]

# Testing
pytest
pytest-asyncio
httpx
faker